backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
databases==0.9.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==5.0.0

# AI APIs
google-generativeai==0.3.2
groq==0.36.0
requests==2.32.3

# Document Generation
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.1.0

# Google APIs
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.3.0
